Add files via upload
System In C Programming language 
Updated an hour ago 
This Project was done by my self in C Programming Language. College Project 1 Project completed in 5 version(Milestones):
Final version(Milestone 5): IPCMS5
Project Commencement Date : OCTOBER, 2016 Project Completion Date : 2nd December, 2016
PROJECT INFORMATION:
This application helps keep the track of an item or product available in a grocery store. This application will show in how much quantity a product or quantity is left. If quantity drops, this application will generate a warning about quantity dropping. It will also show the price of an Item or product. This project is about an application that manages the inventory of items needed for a grocery store. This application is able keep track of all the following information about an item:
1- The SKU number 
2- The name (maximum of 20 chars) 
3- Quantity (On hand quantity currently available in the inventory) 
4- Minimum Quantity 
5- Price of the item 
6- Is the item Taxed